youtube_fetcher/requirements.txt
fastapi
uvicorn
sqlalchemy
aiohttp
alembic
python-dotenv
pydantic
pymysql
